bg_pmove: use conventional indentation to fix warnings
gcc 6 warns by default about indentation that could mislead a reader
into misinterpreting what the code does. This instance is caused by
mixing tab-indentation with space-indentation, and would be misleading
if tabs are considered to be worth any number of spaces not equal to 2.
Standardize on indenting with hard tabs, since this is the prevailing
style in nearby code.