Update expr.c
The mixed indentation and 'break after expression' formatting in this code makes it a bit of a landmine, and is causing a large brick of warnings with various compilers and other tools. Adding explicit bracketing and reformatting cases to make code path more explicit, as well as fixing spaces-vs-tab indentation issues due to having to re-indent this function.