files: Write auto-downloaded PK3 files to a downloads/ directory
They are now only searched if a server uses them, and their expected
checksums are included in the filename (the same as the fallback
filename that was previously used if a collision was detected).
Advanced users can copy them from downloads/ into their
expected places.

The filename ends with ".noexec.pk3" so that we can detect which files
were auto-downloaded, even if they are moved into place.

Adding "downloads/", the checksum and ".noexec" makes the lists of
files somewhat larger, so expand the buffers they're written into
to compensate.

Signed-off-by: Simon McVittie <smcv@debian.org>