Makefile: make the openGL backend selectable at build-time
ioquake3 was linking directly to libGL; but in some cases
it is desiderable to link to other openGL implementations
(i.e. you are building for an embedded system, or the platform has
diverse openGL providers).

Now it is possible to select a different library using the
Makefile variable OPENGL_PROVIDER.

If not set, build system falls back to libGL (as it was before).

Signed-off-by: Francesco Giancane <francescogiancane8@gmail.com>