Move all the Altivec code to its own source files.
This means we can optionally use GCC's -maltivec on just these files, which
are chosen at runtime if the CPU supports Altivec, and compile the rest
without it, making a single binary that has both Altivec optimizations but
can still work on G3.

This commit doesn't hook up the magic that would need in the Makefile, though.
Someone else will have to do that (possibly with an option to still build the
rest with -maltivec anyhow, for use on machines you _know_ will have a vector
unit.