Fix issue #379 (Mac Mouse Wheel) by updating SDL to v2.0.10 stable
I compiled SDL 2.0.10 stable from the official sources and replaced the outdated SDL in the ioq3 project. The mouse wheel bug affecting Mac ioq3 is fixed in later versions of SDL2. I compiled ioq3 for Mac using the new libs current ioq3 source and the problem is fixed.

This updates the Mac libSDL2-2.0.0.dylib and libSDL2main.a in /code/libs/macosx, but does not include header update or other operating system library updates so there is no risk of regressions for other operating systems.