Multiplayer in Emscripten builds!
This relies on the HumbleNet library to emulate BSD sockets over WebRTC DataChannel. Actual unreliable UDP packets are used so performance should be similar to native. WebRTC connection establishment requires a signaling server which is built separately.

The dedicated server binary can now be built in WASM and run in the browser. It should also be possible to build a native dedicated server (or client) with USE_HUMBLENET and connect it to browser clients over WebRTC via the same HumbleNet signaling server. However, this would require importing more code from humblenet and libwebrtc, and has not been tried yet.