Initial SDL3 port.
The initial work was done on YouTube: https://youtu.be/i3yVqWYFbCE

This commit has been rebased and tweaked a little since the video.

The SDL3 headers are in the commit before this one, and the SDL2 header
removal follows one after, so as not to junk up this specific commit and make
sure we don't have a commit where the SDL2 build fails.