improvement(web): rename to `index.html`
This simplifies navigating to the web version.
You don't have to search and click the `ioquake3.html` file.

This is especially useful for the GitHub pages deployment.
The deployment workflows sort of expect `index.html`
to be present.
They usually print `.../ioq3/` as the final URL,
but GitHub pages doesn't allow you to browse the files.
It simply says `404` if there is no `index.html`.
This commit should fix it.