cmake: Allow PRODUCT_VERSION to be overridden at build-time
In packaging systems like Debian's, typically the build will not have
a .git/ directory available; or if it does, the commit ID will be a
downstream packaging commit that means nothing to upstream developers.
However, the packaging can inject a more meaningful version number if
given a way to do so.

Signed-off-by: Simon McVittie <smcv@debian.org>