qcommon: Build generically for otherwise unknown architectures
Portability by enumerating every architecture that we know about doesn't
scale very well, especially on Linux, because new architectures keep
being added and some OS distributions aim to support all of them.
Before commit b512b786 "Move Linux ARCH_STRING definition out of
Makefile", the Makefile could make up an arbitrary name for any new
architecture (in practice `uname -m` on Linux) and it would work, as
long as the engine and any native-code mods in use agreed on what that
name would be; but the addition of a second supported build system made
that more difficult to accommodate.

Now that USE_ARCHLESS_FILENAMES is available, if we build with that
option, the only use of the architecture name is to emit it in two
diagnostic messages, so it doesn't seem like there is a need to enforce
the architecture taking one of a small number of known values.

This returns to the state we had after 2a71948f "build: define ARCH_STRING
in Makefile on Linux and other GNU platforms", where ioquake3 can in
principle be built for any architecture, without needing a series of
changes like ebb69f69 to teach it about each one.

We need to explicitly set USE_ARCHLESS_FILENAMES during
architecture-detection, because add_compile_definitions() doesn't seem
to affect try_compile().

Signed-off-by: Simon McVittie <smcv@debian.org>