fix: shotgun not gibbing unless aiming at feet
Closes https://github.com/ioquake/ioq3/issues/794.

Note that now when shooting at two players
standing behind each other,
the second target will take less damage,
because the dead body of the first player will absorb the pellets
until it gets gibbed (that is, up to 4 pellets,
see `GIB_HEALTH` and `DEFAULT_SHOTGUN_DAMAGE`).

The purpose of the removed line is not entirely clear.
Maybe it's just leftover code,
because the height of a dead body is handled in `PM_CheckDuck`.

See the issue description for more details.