fix: increase `s_alDopplerFactor` 1 -> 5
Partially addresses https://github.com/ioquake/ioq3/issues/802.

This makes the Doppler effect with `s_useOpenAL 1`
sound more like that of the original sound library.
With the value of 1 it's basically unnoticeable.
It's still not quite the same, i.e. it's not as powerful
when a rocket flies very close by, but it's better.
Setting a value higher than 5 makes rockets
sound a little more intense and closer to the original,
but then plasmagun gets perhaps too loud,
and also the sound looping becomes too obvious.